+++ test/src/Game/Werewolf/Test/Game.hs
@@ -8,14 +8,14 @@ Maintainer  : public@hjwylde.com
 {-# OPTIONS_HADDOCK hide, prune #-}
 
 module Game.Werewolf.Test.Game (
-    prop_newGameStartsWithSunsetStage,
-    prop_newGameStartsWithSeesEmpty, prop_newGameStartsWithVotesEmpty,
-    prop_newGameUsesGivenPlayers,
+    prop_newGameStartsWithSunsetStage, prop_newGameStartsWithNoSee,
+    prop_newGameStartsWithVotesEmpty, prop_newGameUsesGivenPlayers,
 ) where
 
 import Control.Lens
 
-import qualified Data.Map as Map
+import qualified Data.Map   as Map
+import           Data.Maybe
 
 import Game.Werewolf.Game
 import Game.Werewolf.Player
@@ -23,8 +23,8 @@ import Game.Werewolf.Player
 prop_newGameStartsWithSunsetStage :: [Player] -> Bool
 prop_newGameStartsWithSunsetStage players = isSunset (newGame players)
 
-prop_newGameStartsWithSeesEmpty :: [Player] -> Bool
-prop_newGameStartsWithSeesEmpty players = Map.null $ newGame players ^. sees
+prop_newGameStartsWithNoSee :: [Player] -> Bool
+prop_newGameStartsWithNoSee players = isNothing $ newGame players ^. see
 
 prop_newGameStartsWithVotesEmpty :: [Player] -> Bool
 prop_newGameStartsWithVotesEmpty players = Map.null $ newGame players ^. votes